[INLONG-2523][Improve][Audit] Modify package name according to specification
