+++ inlong-audit/audit-store/src/test/resources/application.properties
@@ -43,7 +43,7 @@ spring.datasource.druid.max-pool-prepared-statement-per-connection-size=20
 
 # mybatis config
 #mybatis.mapper-locations=classpath*:mapper/*.xml
-mybatis.type-aliases-package=org.apache.inlong.store.entities
+mybatis.type-aliases-package=org.apache.inlong.audit.db.entities
 
 # audit config
 audit.config.file.check.enable=false
